cv_model/requirements.txt
ultralytics
opencv-python
Pillow
wandb
roboflow
albumentations